refactor: implementation modules
